refactor: carValidate() 메서드 이동
자동차의 이름을 검증하는 메서드인 RacingService 클래스의 carValiate() 메서드를 더 적절한 역할을 가진 RacingInitService 클래스로 옮겼습니다.
